Create 2.Insert_at_bottom.cpp
In this question we have to insert an element x at the bottom of an stack. So we first will check if the stack is empty or not.
If not then we pop the elements until it is fully empty.
After that we will insert the element x (given in question to be inserted) then we will again push back all the elements back to its position & return the stack.